project-21-scanned-invoice-reading-agent/requirements.txt
langchain-google-genai
langgraph
pypdf
python-dotenv
Pillow